create query-repo all module

+++ src/news/news.service.ts
@@ -1,60 +1,25 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
-import { EntityManager, In } from 'typeorm';
+import {
+  Injectable,
+  InternalServerErrorException,
+  NotFoundException,
+} from '@nestjs/common';
 import { News } from './domain/news.entity';
-import { CreateNewsInput } from './app/input/news-create.input';
+import { CreateNewsInput } from './api/input/news-create.input';
 import { AwsService } from '../aws/aws.service';
-import { UpdateNewsInput } from './app/input/news-update.input';
-import { Category } from '../category/domain/category.entity';
-import { Tag } from '../tag/domain/tag.entity';
+import { UpdateNewsInput } from './api/input/news-update.input';
 import { extractFileName } from '../helpers';
+import { NewsRepo } from './infrastructure/news.repo';
+import { TagQueryRepo } from '../tag/infrastructure/tag.query-repo';
+import { CategoryQueryRepo } from '../category/infrastructure/category.query-repo';
 
 @Injectable()
 export class NewsService {
   constructor(
-    protected readonly em: EntityManager,
     protected readonly aws: AwsService,
+    protected readonly newsRepo: NewsRepo,
+    protected readonly tagQueryRepo: TagQueryRepo,
+    protected readonly categoryQueryRepo: CategoryQueryRepo,
   ) {}
-
-  async getNews() {
-    return this.em.find(News, { relations: ['tags', 'categories'] });
-  }
-
-  async getNewsById(id: number) {
-    const news = await this.em.findOne(News, {
-      where: { id: id },
-      relations: ['tags', 'categories'],
-    });
-
-    if (!news) {
-      throw new NotFoundException(`News with ID: ${id} not found`);
-    }
-
-    await this.em.increment(News, { id: id }, 'viewCount', 1);
-
-    const categoryIds = news.categories.map((category) => category.id);
-
-    const relatedNews = await this.em
-      .createQueryBuilder(News, 'relatedNews')
-      .leftJoinAndSelect('relatedNews.categories', 'category')
-      .where('category.id IN (:...categoryIds)', { categoryIds })
-      .andWhere('relatedNews.id != :newsId', { newsId: id })
-      .orderBy('relatedNews.publicationDate', 'DESC')
-      .take(2)
-      .getMany();
-
-    return {
-      news,
-      relatedNews,
-    };
-  }
-
-  async getNewsByTag(tagId: number) {
-    return this.em
-      .createQueryBuilder(News, 'news')
-      .leftJoinAndSelect('news.tags', 'tag')
-      .where('tag.id = :tagId', { tagId: tagId })
-      .getMany();
-  }
 
   async createNews(input: CreateNewsInput) {
     const image = await this.aws.uploadFile(await input.image);
@@ -64,15 +29,13 @@ export class NewsService {
     newNews.publicationDate = new Date();
     newNews.image = image;
     newNews.sources = input.sources;
-    newNews.tags = await this.em.find(Tag, { where: { id: In(input.tags) } });
-    newNews.categories = await this.em.find(Category, {
-      where: { id: In(input.categories) },
-    });
-    return await this.em.save(newNews);
+    newNews.tags = await this.tagQueryRepo.find(input.tags);
+    newNews.categories = await this.categoryQueryRepo.find(input.categories);
+    return await this.newsRepo.save(newNews);
   }
 
   async updateNews(id: number, input: UpdateNewsInput) {
-    const news = await this.em.findOne(News, { where: { id: id } });
+    const news = await this.newsRepo.findOne(id);
 
     if (!news) {
       throw new NotFoundException(`News with ID: ${id} not found`);
@@ -83,23 +46,25 @@ export class NewsService {
     news.text = input.text;
     news.image = image;
     news.sources = input.sources;
-    news.tags = await this.em.find(Tag, { where: { id: In(input.tags) } });
-    news.categories = await this.em.find(Category, {
-      where: { id: In(input.categories) },
-    });
+    news.tags = await this.tagQueryRepo.find(input.tags);
+    news.categories = await this.categoryQueryRepo.find(input.categories);
 
-    return this.em.save(news);
+    return this.newsRepo.save(news);
   }
 
   async deleteNews(id: number) {
-    const news = await this.em.findOne(News, { where: { id: id } });
+    const news = await this.newsRepo.findOne(id);
     if (!news) {
       throw new NotFoundException(`News with ID: ${id} not found`);
     }
     const fileName = extractFileName(news.image);
-
+    const deleteResult = await this.newsRepo.delete(id);
+    if (deleteResult.affected === 0) {
+      throw new InternalServerErrorException(
+        `Failed to delete news with ID: ${id}`,
+      );
+    }
     await this.aws.deleteImage(fileName);
-    await this.em.delete(News, { where: { id: id } });
     return news;
   }
 }
